refactor 'common' header methods into abstract base
